ET-4922 more sql utils [5/n] (#1055)
* ET-4756-5-more-pg-utils

* improved push_tuple supporting nested tuples, chunking and error handling

* use u32 instead of usize for snippet_idx

* added SqlBitCastU32 helper

* use IterAsTuple::chunks more and chunks.len() => chunks.element_count()

* review change request

+++ web-api/src/personalization/routes.rs
@@ -596,7 +596,7 @@ impl UnvalidatedSemanticSearchRequest {
 #[serde(untagged)]
 enum UnvalidatedSnippetOrDocumentId {
     DocumentId(String),
-    SnippetId { document_id: String, sub_id: usize },
+    SnippetId { document_id: String, sub_id: u32 },
 }
 
 impl From<SnippetOrDocumentId> for InputDocument {
